Create server folder
