+++ src/main/java/ch/heigvd/dai/commands/Server.java
@@ -30,6 +30,8 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import picocli.CommandLine;
 
+import ch.heigvd.dai.server.ClientHandler;
+
 @CommandLine.Command(
     name = "server",
     description = "Launch the server.",
@@ -40,7 +42,6 @@ public class Server implements Callable<Integer> {
 
   private static final int PORT = 2000;
   private static final int SERVER_ID = (int) (Math.random() * 1000000);
-  private static final String TEXTUAL_DATA = "👋 from Server " + SERVER_ID;
 
   @Override
   public Integer call() {
@@ -51,61 +52,12 @@ public class Server implements Callable<Integer> {
 
       while (!serverSocket.isClosed()) {
         Socket clientSocket = serverSocket.accept();
-        executor.submit(new ClientHandler(clientSocket));
+        executor.submit(new ClientHandler(clientSocket, SERVER_ID));
       }
     } catch (IOException e) {
       System.out.println("[Server " + SERVER_ID + "] exception: " + e);
     }
 
     return 0;
-  }
-
-  static class ClientHandler implements Runnable {
-
-    private final Socket socket;
-
-    public ClientHandler(Socket socket) {
-      this.socket = socket;
-    }
-
-    @Override
-    public void run() {
-      try (socket; // This allow to use try-with-resources with the socket
-          BufferedReader in =
-              new BufferedReader(
-                  new InputStreamReader(socket.getInputStream(), StandardCharsets.UTF_8));
-          BufferedWriter out =
-              new BufferedWriter(
-                  new OutputStreamWriter(socket.getOutputStream(), StandardCharsets.UTF_8))) {
-        System.out.println(
-            "[Server "
-                + SERVER_ID
-                + "] new client connected from "
-                + socket.getInetAddress().getHostAddress()
-                + ":"
-                + socket.getPort());
-
-        System.out.println(
-            "[Server " + SERVER_ID + "] received textual data from client: " + in.readLine());
-
-        try {
-          System.out.println(
-              "[Server " + SERVER_ID + "] sleeping for 10 seconds to simulate a long operation");
-          Thread.sleep(10000);
-        } catch (InterruptedException e) {
-          Thread.currentThread().interrupt();
-        }
-
-        System.out.println(
-            "[Server " + SERVER_ID + "] sending response to client: " + TEXTUAL_DATA);
-
-        out.write(TEXTUAL_DATA + "\n");
-        out.flush();
-
-        System.out.println("[Server " + SERVER_ID + "] closing connection");
-      } catch (IOException e) {
-        System.out.println("[Server " + SERVER_ID + "] exception: " + e);
-      }
-    }
   }
 }
